requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv
groq>=1.0.0
